requirements.txt
# requirements.txt
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
chromadb==0.5.5
sentence-transformers==3.0.1
